Fixed sending data to IoT

+++ EnvironmentService/Dtos/IndoorEnvironmentSettingsUpdateDto.cs
@@ -2,8 +2,8 @@ namespace EnvironmentService.Dtos;
 public class IndoorEnvironmentSettingsUpdateDto
 {
     public int Co2Limit { get; set; }
-    public int TemperatureLimit { get; set; }
-    public int HumidityLimit { get; set; }
-    public int LightLimit { get; set; }
+    public float TemperatureLimit { get; set; }
+    public float HumidityLimit { get; set; }
+    public float LightLimit { get; set; }
     public bool LightOn { get; set; }
 }
